requirements.txt
disnake
colorama
requests
schedule
aiohttp
boticordpy
asyncio
pyfiglet
psutil
pymongo
python-dateutil
pynacl
pytz